+++ CHANGELOG.md
@@ -18,4 +18,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Fixed
 
 - Reapply stone tablet crash fix that was mistakenly reverted during the mc1.16->mc1.18 move.
+- Fix some player decoy behavior when entering editmode while flying.
 - Machine guis with a GO-button can now be started with the keypad enter key.